Реорганизация и рефакторинг
